+++ src/com/hanabi/scenes/Scene.java
@@ -1,0 +1,8 @@
+package src.com.hanabi.scenes;
+
+import java.awt.Graphics2D;
+
+public interface Scene {
+    void update();
+    void render(Graphics2D g2d);
+}