+++ CHANGELOG.md
@@ -2,7 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
-## [1.1.0] - 2026-04-28
+## [1.2.0]a - 2026-4-29
+### Added
+- Added Particle system.
+- Entities and Enemies were defined properly.
+- Scenes were introduced to play different schemes.
+- Interactive collision and reaction have been added.
+
+## [1.1.0]a - 2026-04-28
 ### Added
 - Added Rocket's System for Thrust and Velocity
 - Added fixed background for further additions.
@@ -13,7 +20,7 @@ All notable changes to this project will be documented in this file.
 - Memory leak when handling large peer-to-peer file transfers.
 - Resolved naming conflict in the build script.
 
-## [1.0.0] - 2026-03-15
+## [1.0.0]a - 2026-03-15
 ### Added
 - Initial alpha release of Hanabi.
 - Basic directory structure and project configuration.